feat(matches): mover lógica de notificação de matches para tarefa assíncrona e remover código duplicado
