+++ API/users/tasks.py
@@ -1,12 +1,34 @@
 import cloudinary.uploader
 from celery import shared_task
+from django.core.mail import send_mail
+from django.template.loader import render_to_string
+from django.utils.html import strip_tags
 
-from .match import find_and_notify_matches
 from .models import Item, ItemImage
 
 
 @shared_task
+def send_match_notification(to_email, item_name, matches):
+    subject = f"Possíveis matches para o seu item perdido: {item_name}"
+    html_message = render_to_string(
+        "emails/match_notification.html",
+        {"item_name": item_name, "matches": matches},
+    )
+    plain_message = strip_tags(html_message)
+
+    send_mail(
+        subject,
+        plain_message,
+        "acheiunb2024@gmail.com",
+        [to_email],
+        html_message=html_message,
+    )
+
+
+@shared_task
 def find_and_notify_matches_task(target_item_id, max_distance=2):
+    from .match import find_and_notify_matches  # Importação atrasada para evitar ciclo
+
     """Task assíncrona para encontrar e notificar matches."""
     try:
         target_item = Item.objects.get(id=target_item_id)
