+++ API/users/templates/emails/welcome.html
@@ -1,0 +1,27 @@
+<!DOCTYPE html>
+<html lang="pt-br">
+<head>
+    <meta charset="UTF-8">
+    <title>Bem-vindo ao AcheiUnB</title>
+</head>
+<body>
+    <h1>Bem-vindo ao AcheiUnB, {{ name }}!</h1>
+    <p>
+        Estamos felizes por você estar aqui! O AcheiUnB foi criado para ajudar
+        estudantes a encontrar e recuperar itens perdidos de forma rápida e organizada.
+    </p>
+    <h2>Como começar?</h2>
+    <ul>
+        <li><strong>Cadastrar itens:</strong> Adicione informações sobre os itens perdidos ou encontrados.</li>
+        <li><strong>Buscar itens:</strong> Use a barra de pesquisa para procurar itens na plataforma.</li>
+        <li><strong>Entrar em contato:</strong> Converse com outros usuários diretamente pela plataforma.</li>
+    </ul>
+    <p>
+        Caso tenha dúvidas ou sugestões, não hesite em nos contatar.
+    </p>
+    <p>
+        Boa sorte e conte com a gente!<br>
+        Equipe AcheiUnB
+    </p>
+</body>
+</html>
